requirements.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dnspython==2.1.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
idna==2.10
mongoengine==0.23.1
oauthlib==3.1.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pymongo==3.11.4
PySocks==1.7.1
python-dotenv==0.18.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
telemongo==0.1.0
Telethon==1.22.0
toml==0.10.2
./tweepy-4.0.0a0.whl
urllib3==1.26.6
